# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RoblesGPT API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  clerkUserId: string;
  /** @nullable */
  displayName?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  /**
   * @minLength 1
   * @maxLength 50
   */
  displayName: string;
}

export interface Conversation {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export type ConversationList = Conversation[];

export interface ConversationInput {
  /**
   * @minLength 1
   * @maxLength 100
   */
  title: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  createdAt: string;
}

export interface ConversationDetail {
  id: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: ChatMessage[];
}

export interface ChatInput {
  /** @minLength 1 */
  content: string;
  /** @nullable */
  displayName?: string | null;
}

export interface ChatReply {
  userMessage: ChatMessage;
  assistantMessage: ChatMessage;
  conversationTitle: string;
}

export interface ApiError {
  error: string;
}