# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RoblesGPT API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    put:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    UserProfile:
      type: object
      properties:
        clerkUserId:
          type: string
        displayName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [clerkUserId, createdAt]
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 50
      required: [displayName]
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]